requirements.txt
langchain
langchain-community
langchain-groq
sentence-transformers
faiss-cpu
pypdf
streamlit
python-dotenv
groq
pymilvus
fastapi
uvicorn